adding casedata overflow
