+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/domain/SscsCaseDataOverflow.java
@@ -1,0 +1,19 @@
+package uk.gov.hmcts.reform.sscs.ccd.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Data
+@Builder(toBuilder = true)
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@AllArgsConstructor
+@NoArgsConstructor
+public class SscsCaseDataOverflow {
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private YesNo selectNextHmcHearingType;
+}
